refactor: move functions for work with best time data to own module manage_best_time.js
